Refactoring Launcher
